ai-services/requirements.txt
# Core dependencies
fastapi==0.100.0
uvicorn[standard]==0.21.1
python-multipart==0.0.6
python-dotenv==1.0.0

# Document processing
PyMuPDF==1.21.1
python-docx==0.8.11
beautifulsoup4==4.12.2
lxml==4.9.2
requests==2.28.2
aiohttp==3.8.4
aiofiles==22.1.0

# ML and embeddings
sentence-transformers==2.2.2
huggingface_hub==0.12.1
torch==2.0.0
transformers==4.28.1
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.2.2
nltk==3.8.1

# Utilities
tqdm==4.65.0
loguru==0.7.0
psutil==5.9.5
pydantic
pydantic-core
pydantic-settings

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.0

# Development
black==23.3.0
isort==5.12.0
flake8==6.0.0
mypy==1.2.0
pylint==2.17.2
